# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ItemForge API
  version: 0.1.0
  description: >
    HTTP contract between the educator UI and the ItemForge API. The Python AI
    engine is a private dependency of this service and is not part of this
    document; clients never call it directly.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development
tags:
  - name: Health
  - name: Sources
  - name: Questions
  - name: Review
  - name: Metrics
  - name: Export

paths:
  /api/health:
    get:
      tags: [Health]
      summary: Liveness and dependency status
      responses:
        '200':
          description: The API can serve traffic. `status` may be `degraded` if the AI engine is down.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '503':
          description: The database is unreachable.

  /api/sources:
    get:
      tags: [Sources]
      summary: List uploaded sources
      responses:
        '200':
          description: Sources, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Source'
    post:
      tags: [Sources]
      summary: Upload a source document as plain text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSource'
      responses:
        '201':
          description: Source stored. Chunking happens at generation time.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '422':
          $ref: '#/components/responses/ValidationError'

  /api/sources/{id}:
    get:
      tags: [Sources]
      summary: Fetch a source
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Sources]
      summary: Delete a source and its chunks
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Deleted.
        '404':
          $ref: '#/components/responses/NotFound'

  /api/sources/{id}/generate:
    post:
      tags: [Sources]
      summary: Draft questions from a source via the AI engine
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '201':
          description: A generation run and the drafts it produced, already screened.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/UpstreamError'

  /api/questions:
    get:
      tags: [Questions]
      summary: List questions, optionally filtered by status or source
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/QuestionStatus'
        - name: sourceId
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionPage'
    post:
      tags: [Questions]
      summary: Manually author a question (no AI involved)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestion'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '422':
          $ref: '#/components/responses/ValidationError'

  /api/questions/{id}:
    get:
      tags: [Questions]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Questions]
      summary: Edit stem, choices, or metadata. Records an `edited` review event when the stem changes.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestion'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Questions]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Deleted.
        '404':
          $ref: '#/components/responses/NotFound'

  /api/questions/{id}/events:
    get:
      tags: [Review]
      summary: Audit trail for a single question
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewEvent'

  /api/questions/{id}/approve:
    post:
      tags: [Review]
      summary: Accept a draft into the bank
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewAction'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/questions/{id}/reject:
    post:
      tags: [Review]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewAction'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/questions/{id}/reopen:
    post:
      tags: [Review]
      summary: Move an approved or rejected item back into the review queue
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewAction'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/metrics:
    get:
      tags: [Metrics]
      summary: Aggregate Human-AI interaction metrics derived from review events
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metrics'

  /api/export/gift:
    get:
      tags: [Export]
      summary: Export approved questions as Moodle GIFT
      parameters:
        - name: sourceId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: GIFT document
          content:
            text/plain:
              schema:
                type: string

components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request payload failed schema validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The requested state transition is not legal from the current status.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UpstreamError:
      description: The AI engine failed or is unreachable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Health:
      type: object
      required: [status, dialect, dependencies]
      properties:
        status:
          type: string
          enum: [ok, degraded, down]
        version:
          type: string
        dialect:
          type: string
          enum: [sqlite, postgres]
        dependencies:
          type: object
          additionalProperties:
            type: object
            properties:
              status:
                type: string
              detail:
                type: string

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}

    CreateSource:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          maxLength: 300
        courseCode:
          type: string
          maxLength: 32
          nullable: true
        content:
          type: string
          minLength: 40

    Source:
      type: object
      required: [id, title, content, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        courseCode:
          type: string
          nullable: true
        content:
          type: string
        status:
          type: string
          enum: [pending, chunked, failed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    GenerateRequest:
      type: object
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 20
          default: 6
        type:
          type: string
          enum: [multiple_choice, true_false, short_answer]
          default: multiple_choice
        difficulty:
          type: string
          enum: [easy, medium, hard]

    GenerateResponse:
      type: object
      required: [run, questions]
      properties:
        run:
          $ref: '#/components/schemas/GenerationRun'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'

    GenerationRun:
      type: object
      required: [id, sourceId, provider, model, promptVersion, requestedCount, status]
      properties:
        id:
          type: string
          format: uuid
        sourceId:
          type: string
          format: uuid
        provider:
          type: string
        model:
          type: string
        promptVersion:
          type: string
        requestedCount:
          type: integer
        generatedCount:
          type: integer
        acceptedByRubricCount:
          type: integer
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        costUsd:
          type: number
        latencyMs:
          type: integer
        status:
          type: string
          enum: [running, succeeded, failed]
        error:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    Choice:
      type: object
      required: [label, content, isCorrect, ordinal]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        content:
          type: string
        isCorrect:
          type: boolean
        rationale:
          type: string
          nullable: true
        ordinal:
          type: integer

    RubricFinding:
      type: object
      required: [code, severity, message]
      properties:
        code:
          type: string
        severity:
          type: string
          enum: [error, warning]
        message:
          type: string

    QuestionStatus:
      type: string
      enum: [draft, needs_review, approved, rejected]

    Question:
      type: object
      required: [id, stem, type, status, choices, rubricFindings]
      properties:
        id:
          type: string
          format: uuid
        sourceId:
          type: string
          format: uuid
          nullable: true
        generationRunId:
          type: string
          format: uuid
          nullable: true
        stem:
          type: string
        originalStem:
          type: string
        type:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        bloomLevel:
          type: string
          enum: [remember, understand, apply, analyze, evaluate, create]
        explanation:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/QuestionStatus'
        rubricFindings:
          type: array
          items:
            $ref: '#/components/schemas/RubricFinding'
        duplicateOfId:
          type: string
          format: uuid
          nullable: true
        duplicateSimilarity:
          type: number
          nullable: true
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    QuestionPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer

    CreateQuestion:
      type: object
      required: [stem, choices]
      properties:
        sourceId:
          type: string
          format: uuid
          nullable: true
        stem:
          type: string
        type:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        bloomLevel:
          type: string
          enum: [remember, understand, apply, analyze, evaluate, create]
        explanation:
          type: string
          nullable: true
        choices:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/Choice'

    UpdateQuestion:
      type: object
      properties:
        stem:
          type: string
        explanation:
          type: string
          nullable: true
        difficulty:
          type: string
          enum: [easy, medium, hard]
        bloomLevel:
          type: string
          enum: [remember, understand, apply, analyze, evaluate, create]
        type:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'

    ReviewAction:
      type: object
      properties:
        actor:
          type: string
        note:
          type: string
          nullable: true

    ReviewEvent:
      type: object
      required: [id, questionId, action, actor, createdAt]
      properties:
        id:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
        action:
          type: string
          enum: [approved, rejected, edited, reopened]
        actor:
          type: string
        fromStatus:
          type: string
          nullable: true
        toStatus:
          type: string
          nullable: true
        stemBefore:
          type: string
          nullable: true
        stemAfter:
          type: string
          nullable: true
        editDistance:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    Metrics:
      type: object
      required:
        - questionsTotal
        - acceptanceRate
        - medianEditDistance
        - duplicateRate
        - costPerAcceptedUsd
      properties:
        questionsTotal:
          type: integer
        reviewedTotal:
          type: integer
        acceptedTotal:
          type: integer
        rejectedTotal:
          type: integer
        acceptanceRate:
          type: number
          nullable: true
        medianEditDistance:
          type: number
          nullable: true
        duplicateRate:
          type: number
          nullable: true
        costPerAcceptedUsd:
          type: number
          nullable: true
        rubricErrorRate:
          type: number
          nullable: true